+++ src/test/java/happynewmoonwithreport/type/WasmStringTest.java
@@ -37,7 +37,7 @@ public class WasmStringTest {
         byte[] fox = new byte[]{0x54, (byte) 0x68, (byte) 0x65, (byte) 0x20, (byte) 0x71, (byte) 0x75, (byte) 0x69,
                 (byte) 0x63, (byte) 0x6B, (byte) 0x20, (byte) 0x62, (byte) 0x72, (byte) 0x6F, (byte) 0x77, (byte) 0x6E};
         BytesFile payload = new BytesFile(fox);
-        WasmString wasmFox = new WasmString(payload, new UInt8(fox.length));
+        WasmString wasmFox = new WasmString(payload, new UInt32(fox.length));
         assertEquals(new WasmString("The quick brown"), wasmFox);
 
     }
